ImageLevel/requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
pandas>=1.3.0
pydicom>=2.3.0
scipy>=1.9.0
scikit-image>=0.19.0
cryptography>=41.0.0   # benchmarks/b6 ECDSA comparison only

# Node.js runtime is also required (not a pip package):
#   npm install   (from the ImageLevel/ directory)
# This installs circomlibjs which provides the BN254 Poseidon hash
# used by poseidon_helper.js – the same implementation the circuit uses.

